refactor: 문서 작성 페이지 리팩토링
기존 MakeSection 파일에 에디터들을 포함시켰으나 EditorBox라는 새로운
컴포넌트를 생성하여 해당 파일로 이동 그와 연관된 코드들 이동
